Added justifiaction in risk
