+++ opencti-platform/opencti-front/src/private/components/riskAssessment/risks/RiskDetails.js
@@ -2,8 +2,7 @@ import React, { Component } from 'react';
 import * as PropTypes from 'prop-types';
 import { compose } from 'ramda';
 import * as R from 'ramda';
-import * as Yup from 'yup';
-import { Formik, Form, Field } from 'formik';
+import { Field } from 'formik';
 import { createFragmentContainer } from 'react-relay';
 import graphql from 'babel-plugin-relay/macro';
 import { withStyles } from '@material-ui/core/styles';
@@ -15,7 +14,6 @@ import Edit from '@material-ui/icons/Edit';
 import Switch from '@material-ui/core/Switch';
 import { Information } from 'mdi-material-ui';
 import inject18n from '../../../../components/i18n';
-import { commitMutation } from '../../../../relay/environment';
 import MarkDownField from '../../../../components/MarkDownField';
 import RiskStatus from '../../common/form/RiskStatus';
 import DatePickerField from '../../../../components/DatePickerField';
@@ -93,84 +91,20 @@ const styles = (theme) => ({
   },
 });
 
-const riskDetailsEditMutation = graphql`
-  mutation RiskDetailsEditMutation($id: ID!, $input: [EditInput]!) {
-    editRisk(id: $id, input: $input) {
-      id
-      statement
-      deadline
-      risk_status
-      accepted
-      false_positive
-      risk_adjusted
-      vendor_dependency
-    }
-  }
-`;
-
-const RiskValidation = () => Yup.object().shape({
-  statement: Yup.string().nullable(),
-  risk_status: Yup.string().nullable(),
-  deadline: Yup.string().nullable(),
-  false_positive: Yup.string().nullable(),
-  risk_adjusted: Yup.string().nullable(),
-  vendor_dependency: Yup.string().nullable(),
-  accepted: Yup.string().nullable(),
-});
-
 class RiskDetailsComponent extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      open: false,
-      modelName: '',
-    };
-  }
-
-  handleEditOpen(field) {
-    this.setState({ open: !this.state.open, modelName: field });
-  }
-
-  handleSubmitField(name, value) {
-    RiskValidation(this.props.t)
-      .validateAt(name, { [name]: value })
-      .then(() => {
-        commitMutation({
-          mutation: riskDetailsEditMutation,
-          variables: { id: this.props.risk.id, input: { key: name, value } },
-          onCompleted: () => {
-            this.setState({ modelName: '', open: false });
-          },
-        });
-      })
-      .catch(() => false);
-  }
-
   render() {
     const {
       t,
       classes,
       risk,
       fldt,
-    } = this.props;
-    const {
+      handleSubmitField,
+      handleEditOpen,
       open,
       modelName,
-    } = this.state;
+    } = this.props;
     const riskDetectionSource = R.pipe(
       R.path(['origins']),
-    )(risk);
-    const initialValues = R.pipe(
-      R.assoc('deadline', risk?.deadline || ''),
-      R.assoc('statement', risk?.statement || ''),
-      R.assoc('risk_status', risk?.risk_status || ''),
-      R.assoc('vendor_dependency', risk?.vendor_dependency || false),
-      R.pick([
-        'deadline',
-        'statement',
-        'risk_status',
-        'vendor_dependency',
-      ]),
     )(risk);
     return (
       <div>
@@ -178,11 +112,6 @@ class RiskDetailsComponent extends Component {
           {t('Details')}
         </Typography>
         <Paper classes={{ root: classes.paper }} elevation={2}>
-          <Formik
-            enableReinitialize={true}
-            initialValues={initialValues}
-          >
-            <Form>
               <Grid container spacing={3}>
                 <Grid item={true} xs={6}>
                   <div className={classes.textBase}>
@@ -247,7 +176,7 @@ class RiskDetailsComponent extends Component {
                       size='small'
                       style={{ fontSize: '15px' }}
                       color={(open && modelName === 'statement') ? 'primary' : 'inherit'}
-                      onClick={this.handleEditOpen.bind(this, 'statement')}
+                      onClick={handleEditOpen.bind(this, 'statement')}
                     >
                       <Edit fontSize='inherit' />
                     </IconButton>
@@ -260,7 +189,7 @@ class RiskDetailsComponent extends Component {
                       fullWidth={true}
                       multiline={true}
                       variant='outlined'
-                      onSubmit={this.handleSubmitField.bind(this)}
+                      onSubmit={handleSubmitField.bind(this)}
                     />
                   ) : (
                     <div className={classes.scrollBg}>
@@ -293,7 +222,7 @@ class RiskDetailsComponent extends Component {
                       size='small'
                       style={{ fontSize: '15px' }}
                       color={(open && modelName === 'risk_status') ? 'primary' : 'inherit'}
-                      onClick={this.handleEditOpen.bind(this, 'risk_status')}
+                      onClick={handleEditOpen.bind(this, 'risk_status')}
                     >
                       <Edit fontSize='inherit' />
                     </IconButton>
@@ -304,7 +233,7 @@ class RiskDetailsComponent extends Component {
                       variant='outlined'
                       name='risk_status'
                       size='small'
-                      onChange={this.handleSubmitField.bind(this)}
+                      onChange={handleSubmitField.bind(this)}
                       fullWidth={true}
                       style={{ height: '38.09px', marginBottom: '3px' }}
                       containerstyle={{ width: '100%', padding: '0 0 1px 0' }}
@@ -341,7 +270,7 @@ class RiskDetailsComponent extends Component {
                       size='small'
                       style={{ fontSize: '15px' }}
                       color={(open && modelName === 'deadline') ? 'primary' : 'inherit'}
-                      onClick={this.handleEditOpen.bind(this, 'deadline')}
+                      onClick={handleEditOpen.bind(this, 'deadline')}
                     >
                       <Edit fontSize='inherit' />
                     </IconButton>
@@ -354,7 +283,7 @@ class RiskDetailsComponent extends Component {
                       fullWidth={true}
                       multiline={true}
                       variant='outlined'
-                      onSubmit={this.handleSubmitField.bind(this)}
+                      onSubmit={handleSubmitField.bind(this)}
                       invalidDateMessage={t('The value must be a date (YYYY-MM-DD)')}
                     />
                   ) : risk.deadline && fldt(risk.deadline)
@@ -407,7 +336,7 @@ class RiskDetailsComponent extends Component {
                       size='small'
                       style={{ fontSize: '15px' }}
                       color={(open && modelName === 'vendor_dependency') ? 'primary' : 'inherit'}
-                      onClick={this.handleEditOpen.bind(this, 'vendor_dependency')}
+                      onClick={handleEditOpen.bind(this, 'vendor_dependency')}
                     >
                       <Edit fontSize='inherit' />
                     </IconButton>
@@ -421,7 +350,7 @@ class RiskDetailsComponent extends Component {
                         name="vendor_dependency"
                         type='checkbox'
                         containerstyle={{ margin: '0 -15px 0 11px' }}
-                        onChange={this.handleSubmitField.bind(this)}
+                        onChange={handleSubmitField.bind(this)}
                       />
                     ) : (
                       <Switch
@@ -438,8 +367,6 @@ class RiskDetailsComponent extends Component {
                   </div>
                 </Grid>
               </Grid>
-            </Form>
-          </Formik>
         </Paper>
       </div>
     );
@@ -459,13 +386,6 @@ const RiskDetails = createFragmentContainer(
     risk: graphql`
       fragment RiskDetails_risk on Risk {
         id
-        statement
-        risk_status
-        deadline
-        false_positive
-        risk_adjusted
-        accepted
-        vendor_dependency
         impacted_control_id
         first_seen
         last_seen
